+++ LTN/src/symboles/Li.cpp
@@ -7,7 +7,7 @@ Li::Li() : Symbole(I_Li, "I_Li")
 
 Li::~Li()
 {
-	for(int i=0;i<listeInstructions.size();i++)
+	for(int i=0;i<(int)listeInstructions.size();i++)
 		delete listeInstructions[i];
 }
 
@@ -19,7 +19,7 @@ bool Li::setParam(Symbole* s, int noPlace)
 
 bool Li::executer()
 {
-	for(int i=0;i<listeInstructions.size();i++)
+	for(int i=0;i<(int)listeInstructions.size();i++)
 		if(!listeInstructions[i]->executer())
 			return false;
